requirements.txt
backcall==0.1.0
certifi==2018.8.24
chardet==3.0.4
colorama==0.4.0
decorator==4.3.0
Django==1.11.16
django-mysql==2.4.1
djangorestframework==3.8.2
httpie==0.9.9
idna==2.7
ipdb==0.11
ipython==7.0.1
ipython-genutils==0.2.0
jedi==0.13.0
mysqlclient==1.3.10
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.5
ptyprocess==0.6.0
Pygments==2.2.0
pytz==2018.5
requests==2.19.1
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
urllib3==1.23
wcwidth==0.1.7
django-money-0.14.3
py-moneyed-0.7.0
django-countries-5.3.2

